Implemented invoice deletion

+++ php/deleteinvoice.php
@@ -7,24 +7,56 @@
     </head>
 
     <body>
-        <div class="wrapper-center-children">
+        <div id="wrapper-center-children">
             <!-- TODO: Implement database invoice deletion -->
             <div id="header">Delete Invoice Page </div>
             <?PHP
-            //Use unlink to delete invoice file, display error if unable
-            $filepath = "./invoice" . $_POST['invoiceNumDel'] . ".txt";
-            if (unlink($filepath)) {
-                echo '<h2> Invoice Deleted Successfully </h2>';
-            } else {
-                echo '<h2> There was an error  deleting your file </h2>';
+
+            if (isset($_POST['invoiceNumDel'])) {
+                //Use unlink to delete invoice file, display error if unable
+                $filepath = "./invoice" . $_POST['invoiceNumDel'] . ".txt";
+                if (unlink($filepath)) {
+                    echo '<h2> Invoice Deleted Successfully </h2>';
+                } else {
+                    echo '<h2> There was an error  deleting your file </h2>';
+                }
+                echo '<div>
+                    <input class="thanks-button" style="min-width:200px;" type="button"
+                           onclick="location.href = \'viewfileinvoices.php\';"
+                           value="Go Back to Invoice List" />
+                    </br>
+                  </div>';
+            } elseif (isset($_POST['invoice-number'])) {
+              //Database login Info
+              $server = 'localhost';
+              $user = 'eb3465';
+              $pwd = '55452112eb';
+              $db = 'invoicestorage';
+
+              $conn = mysqli_connect($server, $user, $pwd, $db);
+              if (mysqli_connect_errno())
+                {
+                echo "Failed to connect to MySQL: " . mysqli_connect_error();
+                die();
+              } else {
+                 if (mysqli_query($conn, 'DELETE FROM invoices WHERE invoice_id='.$_POST["invoice-number"].'')) {
+                    echo '<h2> Invoice Deleted Successfully </h2>';
+                 } else {
+                   echo '<h2> There was an error  deleting your file </h2>';
+                 }
+               }
+               echo '<div>
+                   <input class="thanks-button" style="min-width:200px;" type="button"
+                          onclick="location.href = \'viewdbinvoices.php\';"
+                          value="Go Back to Invoice List" />
+                   </br>
+                 </div>';
             }
+
             ?>
             <!-- Naviagtion buttons to go back to other pages -->
-            <div>
-                <input class="thanks-button" style="min-width:200px;" type="button"
-                       onclick="location.href = 'viewfileinvoices.php';"
-                       value="Go Back to Invoice List" />
-                </br>
+
+              <div>
                 <input class="thanks-button" style="min-width:150px;" type="button"
                        onclick="location.href = '../index.html';"
                        value="Create New Form" />
